Support entering new word with Enter
Wrap the text input and main button in a <form>, which means that
pressing Enter in the text input will generate a “click” event for the
button, updating the text. Add a preventDefault() call to stop the
default form handling, which would reload the page. Mark all the other
buttons as type="button", with no default behavior at all.